refactor FileDownloader to extract and move to c.g.c.t.e.util bundle
