+++ plugins/com.google.cloud.tools.eclipse.util/META-INF/MANIFEST.MF
@@ -13,6 +13,8 @@ Export-Package: com.google.cloud.tools.eclipse.util,
  com.google.cloud.tools.eclipse.util.status,
  com.google.cloud.tools.eclipse.util.templates.appengine
 Import-Package: com.google.common.base;version="15.0.0",
+ com.google.common.io;version="17.0.0",
+ com.google.common.net;version="17.0.0",
  org.eclipse.core.commands,
  org.eclipse.core.resources,
  org.eclipse.core.runtime;bundle-symbolic-name="org.eclipse.core.runtime",
